nouuouou

+++ philo.h
@@ -6,7 +6,7 @@
 /*   By: lsouquie <lsouquie@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/02 12:35:42 by lsouquie          #+#    #+#             */
-/*   Updated: 2023/08/11 19:19:48 by lsouquie         ###   ########.fr       */
+/*   Updated: 2023/08/16 19:34:16 by lsouquie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,6 +18,7 @@
 # include <stdlib.h>
 # include <sys/time.h>
 # include "struct.h"
+# include <unistd.h>
 
 
 
@@ -25,7 +26,7 @@ int		check_integrity(char **argv);
 void	birth_of_philos(t_data *data);
 void	init_struct(t_data *data, char **argv);
 void 	*philo_routine(void *data);
-int	get_time();
+long long int	get_time();
 
 
 void need_to_eat(void *data);
